requirements.txt

dj-database-url==0.5.0
Django==3.1.2
django-smart-selects==1.5.9
gunicorn==20.0.4
Pillow==7.2.0
psycopg2==2.9.1
whitenoise==5.2.0
urllib3==1.26.2

